requirements.txt
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
django-cors-headers==3.11.0
django-environ==0.8.1

gunicorn==20.1.0

psycopg2-binary==2.9.3
